Address review comment: Move useful utility into shared module.
