+++ flocker/ca/test/test_ca.py
@@ -1,4 +1,4 @@
-# Copyright Copyright ClusterHQ Inc.  See LICENSE file for details.
+# Copyright ClusterHQ Inc.  See LICENSE file for details.
 
 """
 Tests for certification logic in ``flocker.ca._ca``
@@ -18,7 +18,7 @@ from twisted.python.filepath import FilePath
 from .. import (RootCredential, ControlCredential, NodeCredential,
                 UserCredential, PathError, EXPIRY_20_YEARS,
                 AUTHORITY_CERTIFICATE_FILENAME, AUTHORITY_KEY_FILENAME)
-
+from ..testtools import assert_has_extension
 from ...testtools import not_root, skip_on_broken_permissions
 
 NODE_UUID = str(uuid4())
@@ -339,30 +339,6 @@ class NodeCredentialTests(
         """
         assert_has_extension(self, self.credential.credential,
                              b"extendedKeyUsage", b"clientAuth")
-
-
-def assert_has_extension(test, credential, name, value):
-    """
-    Assert that the ``X509Extension`` with the matching name from the
-    certificate has the given value.
-
-    :param TestCase test: The current test.
-    :param FlockerCredential certificate: Credential whose certificate we
-        should inspect.
-    :param bytes name: The name of the extension.
-    :param bytes value: The data encoded in the extension.
-
-    :raises AssertionError: If the extension is not found or has the wrong
-        value.
-    """
-    expected = crypto.X509Extension(name, False, value)
-    x509 = credential.certificate.original
-    for i in range(x509.get_extension_count()):
-        extension = x509.get_extension(i)
-        if extension.get_short_name() == name:
-            test.assertEqual(extension.get_data(), expected.get_data())
-            return
-    test.fail("Couldn't find extension {}.".format(name))
 
 
 class ControlCredentialTests(
